add ability to unpack gzipped tar of gan mn data
unify script names, config, and logging
