+++ .gitignore
@@ -5,8 +5,9 @@ __pycache__/
 .env
 *.log
 .DS_Store
+
 data/globe_at_night/*.csv
-data/gan_mn/*.csv
+data/gan_mn/*
 data/gan_mn.csv
+
 prediction/*.csv
-validation/*.csv
