+++ ctts/model/config.ini
@@ -2,8 +2,9 @@
 label = nsb
 epochs = 10
 
-[build]
+[file]
 gan_mn_dataframe_filename = gan_mn.csv
+model_filename = model.pth
 
 [log]
 level = 20
