+++ ctts/model/constants.py
@@ -1,5 +1,3 @@
 HIDDEN_SIZE = 64 * 3
 OUTPUT_SIZE = 1
 HOURS_IN_DAY = 24.0
-
-MODEL_STATE_DICT_FILE_NAME = "model.pth"
