+++ data/sources.md
@@ -1,9 +1,9 @@
 # data
 
+## globe at night monitoring network
+
+- https://globeatnight.org/gan-mn/
+
 ## artificial night sky brightness
 
 - https://djlorenz.github.io/astronomy/lp2022/
-
-## globe at night
-
-- https://globeatnight.org/gan-mn/
